news-ai/requirements.txt
streamlit
requests
python-dotenv
transformers
torch
